gpt-5-5: move timeline domain model to data layer (#253)
Relocate TimelineEntry and timelineEntries from TimelineSection to
src/data/timeline.ts; wire UI via type/value imports; point resume-audit
at ./timeline. METADATA_LINE_COUNT stays in component (render concern).
Branch based on #252 so architecture.test.ts passes. All 445 tests green.
