+++ src/components/TimelineSection.tsx
@@ -1,5 +1,7 @@
 import { useMemo, useRef } from 'react'
 import { useTypewriter } from '../animations/useTypewriter'
+import type { TimelineEntry } from '../data/timeline'
+import { timelineEntries } from '../data/timeline'
 import { useTimelineScroll } from '../hooks/useTimelineScroll'
 import { getTimelineScrollRangeVh, getTimelineSnapAnchorTopVh } from './timelineGeometry'
 
@@ -8,41 +10,6 @@ const TIMELINE_SECTION_NO = '// 03'
 function formatPanelCount(index: number, total: number): string {
   return `${String(index + 1).padStart(2, '0')} / ${String(total).padStart(2, '0')}`
 }
-
-const timelineEntries: TimelineEntry[] = [
-  {
-    hash: 'd4e8f2c',
-    dateRange: 'AUG 2024 – PRESENT',
-    institution: 'NETAPP INC.',
-    role: 'SOFTWARE_ENGINEER_IN_TEST',
-    bullets: [
-      'Built automated analysis pipeline processing storage telemetry across distributed RAID systems (FC, SAS, NVMe/RoCE), handling terabytes of performance data.',
-      'Designed Python automation framework reducing manual configuration tasks by 30% across Linux, Windows, and VMware infrastructure.',
-      'Implemented Ansible-based deployment orchestration for 300+ system configurations, streamlining infrastructure provisioning workflows.',
-      'Developed interactive visualization dashboard for system performance metrics using Python, analyzing 1M+ database entries for engineering insights.',
-    ],
-    category: 'experience',
-  },
-  {
-    hash: 'a3f9d2b',
-    dateRange: 'JAN 2026 – DEC 2027 (EXPECTED)',
-    institution: 'WICHITA STATE UNIVERSITY',
-    role: 'ACCELERATED_M.S._COMPUTER_SCIENCE',
-    bullets: [],
-    category: 'education',
-  },
-  {
-    hash: 'b7c3e1a',
-    dateRange: 'JAN 2022 – DEC 2025',
-    institution: 'WICHITA STATE UNIVERSITY',
-    role: 'B.S._COMPUTER_SCIENCE',
-    bullets: [
-      "Dean's List: Spring 2022 – Fall 2025",
-      'Relevant Coursework: Machine Learning, Artificial Intelligence, Fundamentals of AI Agents, Data Science',
-    ],
-    category: 'education',
-  },
-]
 
 const METADATA_LINE_COUNT = 3
 
@@ -226,14 +193,3 @@ const TimelineSection: React.FC = () => {
 }
 
 export default TimelineSection
-
-export interface TimelineEntry {
-  hash: string
-  dateRange: string
-  institution: string
-  role: string
-  bullets: string[]
-  category: 'experience' | 'education'
-}
-
-export { timelineEntries }
