+++ src/data/timeline.ts
@@ -1,0 +1,43 @@
+export interface TimelineEntry {
+  hash: string
+  dateRange: string
+  institution: string
+  role: string
+  bullets: string[]
+  category: 'experience' | 'education'
+}
+
+export const timelineEntries: TimelineEntry[] = [
+  {
+    hash: 'd4e8f2c',
+    dateRange: 'AUG 2024 – PRESENT',
+    institution: 'NETAPP INC.',
+    role: 'SOFTWARE_ENGINEER_IN_TEST',
+    bullets: [
+      'Built automated analysis pipeline processing storage telemetry across distributed RAID systems (FC, SAS, NVMe/RoCE), handling terabytes of performance data.',
+      'Designed Python automation framework reducing manual configuration tasks by 30% across Linux, Windows, and VMware infrastructure.',
+      'Implemented Ansible-based deployment orchestration for 300+ system configurations, streamlining infrastructure provisioning workflows.',
+      'Developed interactive visualization dashboard for system performance metrics using Python, analyzing 1M+ database entries for engineering insights.',
+    ],
+    category: 'experience',
+  },
+  {
+    hash: 'a3f9d2b',
+    dateRange: 'JAN 2026 – DEC 2027 (EXPECTED)',
+    institution: 'WICHITA STATE UNIVERSITY',
+    role: 'ACCELERATED_M.S._COMPUTER_SCIENCE',
+    bullets: [],
+    category: 'education',
+  },
+  {
+    hash: 'b7c3e1a',
+    dateRange: 'JAN 2022 – DEC 2025',
+    institution: 'WICHITA STATE UNIVERSITY',
+    role: 'B.S._COMPUTER_SCIENCE',
+    bullets: [
+      "Dean's List: Spring 2022 – Fall 2025",
+      'Relevant Coursework: Machine Learning, Artificial Intelligence, Fundamentals of AI Agents, Data Science',
+    ],
+    category: 'education',
+  },
+]
